requirements.txt
# PyWatchdog - Dependências
# Framework Web
Flask==2.3.3

# Configuração
PyYAML==6.0.1
python-dotenv==1.0.0

# Criptografia & Segurança
pycryptodome==3.19.0

# Requisições HTTP
requests==2.31.0

# Monitoramento de Arquivos
watchdog==3.0.0

# Dependências do Flask
Werkzeug==2.3.7
Jinja2==3.1.2
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3
blinker==1.6.3

# Testes
pytest==7.4.3
pytest-cov==4.1.0

# Desenvolvimento
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Documentação
sphinx==7.2.6
sphinx-rtd-theme==1.3.0